+++ aldrin/src/low_level/event.rs
@@ -1,4 +1,5 @@
 use crate::core::{Deserialize, DeserializeError, SerializedValue, SerializedValueSlice};
+use crate::unknown_event::UnknownEvent;
 
 /// Event emitted by a service.
 #[derive(Debug, Clone)]
@@ -30,5 +31,10 @@ impl Event {
     /// Deserializes the event's arguments.
     pub fn deserialize<T: Deserialize>(&self) -> Result<T, DeserializeError> {
         self.args.deserialize()
+    }
+
+    /// Converts this event into an [`UnknownEvent`].
+    pub fn into_unknown_event(self) -> UnknownEvent {
+        UnknownEvent::new(self)
     }
 }